feat: Add ADSB map refactor: Split Map class to files refactor: move interfaces to type file
