+++ index.html
@@ -37,6 +37,9 @@
         <label for="open-railway">Open railway</label>
         <input type="radio" id="bing" name="map-type" value="bing">
         <label for="bing">BING</label>
+        <input type="radio" id="adsb" name="map-type" value="adsb">
+        <label for="adsb">ADSB</label>
+
     </nav>
 </header>
 <main class="layout">
