--- src/Observer.interface.ts
@@ -1,5 +1,0 @@
-import { Message } from './types';
-
-export interface Observer {
-  update: (publication: Message) => void;
-}
