+++ src/main.ts
@@ -6,7 +6,8 @@ import 'leaflet-geosearch/dist/geosearch.css';
 import { MAX_ZOOM } from './constants.ts';
 import './styles/style.css';
 import { Axis } from './Axis.class.ts';
-import { MapFactory, OsmFrame } from './Map.class.ts';
+import { MapFactory } from './MapFactory.class.ts';
+import { OsmFrame } from './Providers.ts';
 import { RadioGroupClass } from './RadioGroup.class.ts';
 import { Scene } from './Scene.class.ts';
 import { osmLayers } from './layers.ts';
